feat: tc layer service barrel import optimisation
