+++ app/classes/Tokens.ts
@@ -1,11 +1,12 @@
 import { unknownToken } from '../data/token'
+import { tokenCacheApi } from '../service/cache'
+import { sharedTokenClient } from '../service/token'
 import { TokenVerification } from '@injectivelabs/sdk-ts/types'
 /**
  * @deprecated Use subpath imports from '@injectivelabs/sdk-ts/service' instead.
  * TokenFactoryStatic is temporarily imported from the barrel until all apps migrate.
  */
 import { TokenFactoryStatic } from '@injectivelabs/sdk-ts/service'
-import { tokenCacheApi, sharedTokenClient } from '../service'
 import type { Coin, TokenStatic } from '@injectivelabs/sdk-ts'
 import type { SharedBalanceWithToken } from '../types'
