+++ app/service/bff.ts
@@ -1,0 +1,4 @@
+import { BffApi } from './app/bffApi'
+import { ENDPOINTS } from '../utils/constant'
+
+export const bffApi = new BffApi(ENDPOINTS.bffApi)
