+++ app/service/derivative.ts
@@ -1,0 +1,5 @@
+import { ENDPOINTS } from '../utils/constant'
+import { getIndexerGrpcDerivativesApi } from '../utils/lib/sdkImports'
+
+export const getIndexerDerivativesApi = () =>
+  getIndexerGrpcDerivativesApi(ENDPOINTS.indexer)
