+++ app/service/faucet.ts
@@ -1,0 +1,3 @@
+import { FaucetService } from './app/Faucet'
+
+export const faucetService = new FaucetService()
