+++ app/service/name.ts
@@ -1,0 +1,5 @@
+import { InjNameService } from './app/nameService'
+import { InjBonfidaNameService } from './app/bonfida'
+
+export const injNameService = new InjNameService()
+export const injBonfidaNameService = new InjBonfidaNameService()
