+++ app/service/oracle.ts
@@ -1,0 +1,5 @@
+import { ENDPOINTS } from '../utils/constant'
+import { getIndexerGrpcOracleApi } from '../utils/lib/sdkImports'
+
+export const getIndexerOracleApi = () =>
+  getIndexerGrpcOracleApi(ENDPOINTS.indexer)
