+++ app/service/pyth.ts
@@ -1,0 +1,3 @@
+import { PythService } from './app/pythClient'
+
+export const pythService = new PythService()
