+++ app/service/rwa.ts
@@ -1,0 +1,3 @@
+import { CacheRwaPriceFeedService } from './app/cacheRwaPriceFeedService'
+
+export const cacheRwaPriceFeedService = new CacheRwaPriceFeedService()
