+++ app/service/token.ts
@@ -1,0 +1,5 @@
+import { SharedTokenClient } from './app/tokenClient'
+import { SharedTokenClientStatic } from './app/SharedTokenClientStatic'
+
+export const sharedTokenClient = new SharedTokenClient()
+export const sharedTokenClientStatic = new SharedTokenClientStatic()
