+++ app/service/tokenStaticFactory.ts
@@ -1,0 +1,3 @@
+import { TokenStaticFactory } from '@injectivelabs/sdk-ts/service'
+
+export const tokenStaticFactory = new TokenStaticFactory([])
